Created PlaceProvider.
